+++ src/test/java/io/github/eocqrs/cmig/session/SimpleTest.java
@@ -1,0 +1,70 @@
+/*
+ *  Copyright (c) 2023 Aliaksei Bialiauski, EO-CQRS
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NON-INFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+
+package io.github.eocqrs.cmig.session;
+
+import com.datastax.driver.core.Cluster;
+import org.hamcrest.MatcherAssert;
+import org.hamcrest.Matchers;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.Mock;
+import org.mockito.junit.jupiter.MockitoExtension;
+
+/**
+ * Test suite for {@link Simple}.
+ *
+ * @author Aliaksei Bialiauski (abialiauski.dev@gmail.com)
+ * @since 0.0.0
+ */
+@ExtendWith(MockitoExtension.class)
+final class SimpleTest {
+
+  @Test
+  void createsCassandra(@Mock final Cluster mock) {
+    final Cassandra cassandra = new Simple(mock);
+    MatcherAssert.assertThat(
+      "Creates Cassandra",
+      cassandra,
+      Matchers.notNullValue()
+    );
+  }
+
+  @Test
+  void closesWithoutException(@Mock final Cluster mock) {
+    final Cassandra cassandra = new Simple(mock);
+    Assertions.assertDoesNotThrow(
+      cassandra::close,
+      () -> "Closes without exception"
+    );
+  }
+
+  @Test
+  void connectsToMockWithoutException(@Mock final Cluster mock) throws Exception {
+    final Cassandra cassandra = new Simple(mock);
+    Assertions.assertDoesNotThrow(
+      cassandra::value,
+      () -> "Connect to mocked cluster doesnt throw exception"
+    );
+  }
+}
